% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pattern.R
\name{draw_pattern}
\alias{draw_pattern}
\title{Plotting colored multiplication table}
\usage{
draw_pattern(k, xmax, ymax, colors, show_title = TRUE,
  show_numbers = FALSE, show_legend = FALSE)
}
\arguments{
\item{k}{Values for which to find factors}

\item{xmax}{Max value of x to show in the plot}

\item{ymax}{Max value of y to show in the plot}

\item{colors}{Colors to be assigned to factors of k. Defaults to a random palette from wesanderson package}

\item{show_title}{Show title in the plot?}

\item{show_numbers}{Show corresponding numbers in the cells?}

\item{show_legend}{Show legend identifying the color mapping}
}
\description{
Visualize the multiplication table coloring factors of k.
}
