% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Generating the values of the multiplication table}
\usage{
create_table(k, xmax, ymax)
}
\arguments{
\item{k}{Numeric vector specifying the values of k}

\item{xmax}{Max value of x to be generated}

\item{ymax}{Max value of y to be generated}
}
\value{
Data frame
}
\description{
Creates a data frame with the values of the multiplication table and identifies the factors for given k.
}
\examples{
create_table()

}
